requirements.txt
asgiref==3.3.4
autopep8==1.5.7
Django==3.2.4
djangorestframework==3.12.4
numpy==1.19.5
pandas==1.1.5
psycopg2==2.9.1
pycodestyle==2.7.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
tdqm==0.0.1
toml==0.10.2
tqdm==4.61.1
typing-extensions==3.10.0.0
uWSGI==2.0.19.1 ; sys_platform == "linux"
djoser==2.1.0
PyYAML==5.4.1
drf-spectacular==0.17.2
django-extensions==3.1.3